REFACTOR: clean architecture 폴더구조 변경
