+++ lib/core/component/buttons.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 
-import 'package:recipe_app/ui/color_styles.dart';
-import 'package:recipe_app/ui/text_styles.dart';
+import 'package:recipe_app/core/ui_styles/color_styles.dart';
+import 'package:recipe_app/core/ui_styles/text_styles.dart';
 
 enum ButtonSize { big, medium, small }
