+++ lib/core/component/ingredient_item.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:recipe_app/data/model/ingredient/ingredient.dart';
-import 'package:recipe_app/ui/color_styles.dart';
-import 'package:recipe_app/ui/text_styles.dart';
+import 'package:recipe_app/domain/model/ingredient/ingredient.dart';
+import 'package:recipe_app/core/ui_styles/color_styles.dart';
+import 'package:recipe_app/core/ui_styles/text_styles.dart';
 
 class IngredientItem extends StatelessWidget {
   final Ingredient ingredient;
