+++ lib/core/component/skeleton_effect.dart
