+++ lib/data/mepper/ingredient_mepper.dart
@@ -1,5 +1,5 @@
 import 'package:recipe_app/data/dto/recipe_dto.dart';
-import 'package:recipe_app/data/model/ingredient/ingredient.dart';
+import 'package:recipe_app/domain/model/ingredient/ingredient.dart';
 
 extension IngredientMapper on IngredientDto {
   Ingredient toIngredient({int weight = 0}) {
