+++ lib/data/repository/recipe_repository_impl.dart
@@ -1,7 +1,7 @@
 import 'package:recipe_app/data/data_source/recipe_data_source.dart';
 import 'package:recipe_app/data/mepper/recipe_mepper.dart';
-import 'package:recipe_app/data/model/recipe/recipe.dart';
-import 'package:recipe_app/data/repository/recipe_repository.dart';
+import 'package:recipe_app/domain/model/recipe/recipe.dart';
+import 'package:recipe_app/domain/roepositpry/recipe_repository.dart';
 import 'package:recipe_app/presentation/screen/search_recipes/state/filter_search_state.dart';
 
 class RecipeRepositoryImpl implements RecipeRepository {
