+++ lib/domain/model/filter/filter_enum.dart
