+++ lib/domain/roepositpry/recipe_repository.dart
@@ -1,4 +1,4 @@
-import 'package:recipe_app/data/model/recipe/recipe.dart';
+import 'package:recipe_app/domain/model/recipe/recipe.dart';
 
 abstract interface class RecipeRepository {
   Future<List<Recipe>> getRecipes();
