+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:recipe_app/core/routing/router.dart';
-import 'package:recipe_app/ui/color_styles.dart';
+import 'package:recipe_app/core/ui_styles/color_styles.dart';
 
 void main() {
   runApp(const MyApp());
