+++ lib/presentation/example_screen/button_example_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:recipe_app/presentation/component/buttons.dart';
-import 'package:recipe_app/ui/text_styles.dart';
+import 'package:recipe_app/core/component/buttons.dart';
+import 'package:recipe_app/core/ui_styles/text_styles.dart';
 
 class ButtonExampleScreen extends StatelessWidget {
   const ButtonExampleScreen({super.key});
