+++ lib/presentation/example_screen/filter_rating_button_example_screen.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:recipe_app/presentation/component/filter_button.dart';
+import 'package:recipe_app/core/component/filter_button.dart';
 
-import 'package:recipe_app/presentation/component/rating_button.dart';
-import 'package:recipe_app/ui/text_styles.dart';
+import 'package:recipe_app/core/component/rating_button.dart';
+import 'package:recipe_app/core/ui_styles/text_styles.dart';
 
 class FilterRatingButtonExampleScreen extends StatefulWidget {
   const FilterRatingButtonExampleScreen({super.key});
