+++ lib/presentation/example_screen/ingredient_list_example_screen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
-import 'package:recipe_app/data/model/ingredient/ingredient.dart';
-import 'package:recipe_app/presentation/component/ingredient_item.dart';
-import 'package:recipe_app/ui/text_styles.dart';
+import 'package:recipe_app/domain/model/ingredient/ingredient.dart';
+import 'package:recipe_app/core/component/ingredient_item.dart';
+import 'package:recipe_app/core/ui_styles/text_styles.dart';
 
 class IngredientListExampleScreen extends StatelessWidget {
   const IngredientListExampleScreen({super.key});
