+++ lib/presentation/example_screen/input_field_example_screen.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:recipe_app/presentation/component/input_field.dart';
-import 'package:recipe_app/ui/text_styles.dart';
+import 'package:recipe_app/core/component/input_field.dart';
+import 'package:recipe_app/core/ui_styles/text_styles.dart';
 
 class InputFieldExampleScreen extends StatelessWidget {
   const InputFieldExampleScreen({super.key});
