+++ lib/presentation/screen/main_naivation_bar/saved_recipes/saved_recipe_state.dart
@@ -1,6 +1,6 @@
 // ignore_for_file: annotate_overrides
 import 'package:freezed_annotation/freezed_annotation.dart';
-import 'package:recipe_app/data/model/recipe/recipe.dart';
+import 'package:recipe_app/domain/model/recipe/recipe.dart';
 
 part 'saved_recipe_state.freezed.dart';
