+++ lib/presentation/screen/search_recipes/filter_search_bottom_sheet.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
-import 'package:recipe_app/data/model/filter/filter_enum.dart';
-import 'package:recipe_app/presentation/component/buttons.dart';
-import 'package:recipe_app/presentation/component/filter_button.dart';
-import 'package:recipe_app/presentation/component/rating_button.dart';
+import 'package:recipe_app/domain/model/filter/filter_enum.dart';
+import 'package:recipe_app/core/component/buttons.dart';
+import 'package:recipe_app/core/component/filter_button.dart';
+import 'package:recipe_app/core/component/rating_button.dart';
 import 'package:recipe_app/presentation/screen/search_recipes/search_recipes_view_model.dart';
 
-import 'package:recipe_app/ui/text_styles.dart';
+import 'package:recipe_app/core/ui_styles/text_styles.dart';
 
 class FilterSearchBottomSheet extends StatefulWidget {
   final SearchRecipesViewModel viewModel;
