+++ lib/presentation/screen/search_recipes/state/search_recipe_state.dart
@@ -1,7 +1,7 @@
 // ignore_for_file: annotate_overrides
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import 'package:recipe_app/data/model/recipe/recipe.dart';
+import 'package:recipe_app/domain/model/recipe/recipe.dart';
 import 'package:recipe_app/presentation/screen/search_recipes/state/filter_search_state.dart';
 
 part 'search_recipe_state.freezed.dart';
