+++ lib/presentation/screen/sign_in/sign_in_screen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 import 'package:recipe_app/core/routing/routes.dart';
-import 'package:recipe_app/presentation/component/buttons.dart';
-import 'package:recipe_app/presentation/component/input_field.dart';
-import 'package:recipe_app/ui/color_styles.dart';
-import 'package:recipe_app/ui/text_styles.dart';
+import 'package:recipe_app/core/component/buttons.dart';
+import 'package:recipe_app/core/component/input_field.dart';
+import 'package:recipe_app/core/ui_styles/color_styles.dart';
+import 'package:recipe_app/core/ui_styles/text_styles.dart';
 
 class SignInScreen extends StatelessWidget {
   const SignInScreen({super.key});
