+++ test/presentation/component/ingredient_item_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:recipe_app/data/model/ingredient/ingredient.dart';
-import 'package:recipe_app/presentation/component/ingredient_item.dart';
+import 'package:recipe_app/domain/model/ingredient/ingredient.dart';
+import 'package:recipe_app/core/component/ingredient_item.dart';
 
 void main() {
   group('IngredientItemr Button Widget Test', () {
