+++ test/presentation/component/recipe_card_test.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:recipe_app/data/model/filter/filter_enum.dart';
-import 'package:recipe_app/data/model/ingredient/ingredient.dart';
-import 'package:recipe_app/data/model/media/media.dart';
-import 'package:recipe_app/data/model/recipe/recipe.dart';
-import 'package:recipe_app/data/model/user/user.dart';
-import 'package:recipe_app/presentation/component/recipe_card.dart';
+import 'package:recipe_app/domain/model/filter/filter_enum.dart';
+import 'package:recipe_app/domain/model/ingredient/ingredient.dart';
+import 'package:recipe_app/domain/model/media/media.dart';
+import 'package:recipe_app/domain/model/recipe/recipe.dart';
+import 'package:recipe_app/domain/model/user/user.dart';
+import 'package:recipe_app/core/component/recipe_card.dart';
 
 void main() {
   group('RecipeCard Widget Test', () {
