+++ test/presentation/component/tabs_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:recipe_app/presentation/component/tabs.dart';
+import 'package:recipe_app/core/component/tabs.dart';
 
 void main() {
   group('Tabs widget test', () {
